requirements.txt
requests
bs4
pandas
Pillow
io
numpy